pyproject.toml
[project]
name = "hypothesis-council"
version = "0.1.0"
description = "Hypothesis-testing variant of the LLM council pattern"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "tavily-python>=0.5",
    "pydantic>=2.7",
    "click>=8.1",
]

[project.scripts]
council = "council.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
